Upgrades jQuery and RequireJS
Oh my. This is old. This upgrades before mentioned requirements to their
latest available versions.

jQuery changed how to export itself, and as we are using RequireJS, we
needed to specify the same "library name" ("jquery") as the one being
exported by the library. Therefore the changes from "library/jquery" to
"jquery".